core: lift the mask distance transform into a shared header
The exact separable Felzenszwalb-Huttenlocher squared-Euclidean DT and the
dilate/shrink built on it were in DataManager.cpp's anonymous namespace, where
only the dataset decoder could reach them. The SAM masking policy needs the
same dilation to grow a detection before it joins the union, and ss_sam links
ss_nn only -- not the engine -- so the code has to be header-only to be shared.
Moved verbatim into src/core/DistanceTransform.h as namespace edt; the one
call site is now qualified. No behaviour change.

Implementer subagent (Claude Opus). Asked to make the SAM masking in
compute/spirula-studio grow its masks by a configurable margin baked into the
mask PNGs it writes, default 15% and on; this is the first step, sharing the
dilation primitive rather than writing a second one.
